Added Headlines and other Modifications
Changed out the scrollview with the scores to a headline animation.
Users can now add their own headlines in addition to their own
pictures.  Both implemented in the database.  Changed the structure of
the project as well to fit MVC.  Working on the service layer calls
though.  They’re still within the controller layer.